mon: move health_status_t to health.h
